+++ sending/backends/websocket.py
@@ -35,6 +35,9 @@ class WebsocketManager(AbstractPubSubManager):
         super().__init__()
         self.ws_url = ws_url
 
+        # Save / overwrite response headers on each connection
+        self.response_headers = {}
+
         # An unauth_ws and authed_ws pair of Futures are created so that
         # sub-classes can easily implement a pattern where messages are only
         # sent to the server after the session has been authenticated.
@@ -144,6 +147,7 @@ class WebsocketManager(AbstractPubSubManager):
         """
         # Automatic reconnect https://websockets.readthedocs.io/en/stable/reference/client.html
         async for websocket in websockets.connect(self.ws_url):
+            self.response_headers = dict(websocket.response_headers)
             try:
                 self.unauth_ws.set_result(websocket)
                 # Call the auth and init hooks (casting to async if necessary), passing in 'self'